add tensorflow integration
