+++ h2o-algos/build.gradle
@@ -8,6 +8,7 @@ dependencies {
   testCompile "junit:junit:${junitVersion}"
   testCompile project(path: ":h2o-core", configuration: "testArchives")
   compile files('water.gpu.jar')
+  compile files('deepwater.backends.tensorflow-1.0-SNAPSHOT.jar')
 }
 
 apply from: "${rootDir}/gradle/dataCheck.gradle"
