+++ h2o-algos/src/main/java/hex/deepwater/backends/TensorflowBackend.java
@@ -2,56 +2,58 @@ package hex.deepwater.backends;
 
 import hex.deepwater.datasets.ImageDataSet;
 
-class TensorflowBackend implements BackendTrain {
+import deepwater.backends.tensorflow.TensorflowBackendImpl;
 
-    @Override
-    public void delete() {
+class TensorflowBackend extends TensorflowBackendImpl implements BackendTrain {
 
-    }
+//    @Override
+//    public void delete() {
+//        super.delete();
+//    }
 
     @Override
     public void buildNet(ImageDataSet dataset, RuntimeOptions opts, BackendParams backend_params, int num_classes, String name) {
 
     }
 
-    @Override
-    public void saveModel(String model_path) {
+//    @Override
+//    public void saveModel(String model_path) {
+//
+//    }
 
-    }
+//    @Override
+//    public void loadParam(String param_path) {
+//
+//    }
 
-    @Override
-    public void loadParam(String param_path) {
+//    @Override
+//    public void saveParam(String param_path) {
+//
+//    }
 
-    }
+//    @Override
+//    public String toJson() {
+//        return null;
+//    }
 
-    @Override
-    public void saveParam(String param_path) {
+//    @Override
+//    public void setParameter(String name, float value) {
+//
+//    }
 
-    }
+//    @Override
+//    public float[] train(float[] data, float[] label) {
+//        return new float[0];
+//    }
+//
+//    @Override
+//    public float[] predict(float[] data, float[] label) {
+//        return new float[0];
+//    }
 
-    @Override
-    public String toJson() {
-        return null;
-    }
-
-    @Override
-    public void setParameter(String name, float value) {
-
-    }
-
-    @Override
-    public float[] train(float[] data, float[] label) {
-        return new float[0];
-    }
-
-    @Override
-    public float[] predict(float[] data, float[] label) {
-        return new float[0];
-    }
-
-    @Override
-    public float[] predict(float[] data) {
-        return new float[0];
-    }
+//    @Override
+//    public float[] predict(float[] data) {
+//        return new float[0];
+//    }
 
 }
